[Change] 검출코드 통합
NC-192
1) DetectManager + CheckSurrounding 통합
2)DetectManager.CharacterDetect - > CheckSurrounding 의 캐릭터 검출 부분
3) Dir enum -> 위치를 DetectManager로 변경
4) PlayerMovement 의 CheckSurrounding 쓰는 부분 DetectManager로 변경
