+++ Assets/Scripts/DetectObj/DetectManager.CharacterDetect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aaac385ee3aa24e9aa71678a738e1116
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
